# MINOR: [C++][Parquet] Make DeltaBitPack roundtrip tests faster