Before:
```
[----------] 1 test from TestDeltaBitPackEncoding/0, where TypeParam = parquet::PhysicalType<(parquet::Type::type)1>
[ RUN      ] TestDeltaBitPackEncoding/0.BasicRoundTrip
[       OK ] TestDeltaBitPackEncoding/0.BasicRoundTrip (11464 ms)
[----------] 1 test from TestDeltaBitPackEncoding/0 (11464 ms total)

[----------] 1 test from TestDeltaBitPackEncoding/1, where TypeParam = parquet::PhysicalType<(parquet::Type::type)2>
[ RUN      ] TestDeltaBitPackEncoding/1.BasicRoundTrip
[       OK ] TestDeltaBitPackEncoding/1.BasicRoundTrip (12515 ms)
[----------] 1 test from TestDeltaBitPackEncoding/1 (12515 ms total)
```

After:
```
[----------] 1 test from TestDeltaBitPackEncoding/0, where TypeParam = parquet::PhysicalType<(parquet::Type::type)1>
[ RUN      ] TestDeltaBitPackEncoding/0.BasicRoundTrip
[       OK ] TestDeltaBitPackEncoding/0.BasicRoundTrip (259 ms)
[----------] 1 test from TestDeltaBitPackEncoding/0 (259 ms total)

[----------] 1 test from TestDeltaBitPackEncoding/1, where TypeParam = parquet::PhysicalType<(parquet::Type::type)2>
[ RUN      ] TestDeltaBitPackEncoding/1.BasicRoundTrip
[       OK ] TestDeltaBitPackEncoding/1.BasicRoundTrip (242 ms)
[----------] 1 test from TestDeltaBitPackEncoding/1 (242 ms total)
```